zip package ready for manufacturing
First upload of the HW for the blastic scale project.  It is intended for volunteering developers or beta testers who want to get their own hardware materials to test the hardware on their own.  Could be used also by common users, but documentation for best build practices is under construction for now.  Final milestone fot the publishing of that is February 2025, altought it can come sooner.